feat: 플레이어 추상 클래스 생성
